% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPS-package.R
\docType{package}
\name{ALPS-package}
\alias{ALPS}
\alias{ALPS-package}
\title{ALPS: AnaLysis routines for ePigenomicS data}
\description{
ALPS is an R package that provides tools for the analysis and visualization, mainly aimed at epigenomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/itsvenu/ALPS}
  \item Report bugs at \url{https://github.com/itsvenu/ALPS/issues}
}

}
\keyword{internal}
