% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cdf_df.R
\name{gather_griddata}
\alias{gather_griddata}
\title{Read parsed netCDF data for arbitrary variables into dataframes to input
into regression models.}
\usage{
gather_griddata(griddata)
}
\arguments{
\item{geo.ave_data}{The globally averaged data}

\item{gridded_data}{The globally gridded data}
}
\value{
A long form \code{data.frame}.
}
\description{
The variable data should be a pair of parsed netCDF files, one of which should
be globally averaged the other should have globally gridded values.
The data can have any time series (daily, monthly, annually, etc.) as long as
both parsed files have the same length.
}
