% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_regions.R
\name{get_variable_regions}
\alias{get_variable_regions}
\title{Get variable regions}
\usage{
get_variable_regions(enrichments_df = NULL, log_transform = TRUE,
  scale = TRUE, num_regions = 500)
}
\arguments{
\item{enrichments_df}{a data.frame o enrichments at genomic regions. Output of
\code{multiBigwig_summary} or a similar format is compatible}

\item{log_transform}{logical, whether to log2 transform the counts, default \code{TRUE}}

\item{scale}{logical, whether to \code{scale} the variable regions before returning the results, default \code{TRUE}}

\item{num_regions}{number of variable regions to return, default \code{500}}
}
\value{
a data.frame of scaled variable regions
}
\description{
given a data.frame of genomic regions enrichments, the function returns the number of
variable regions across all samples. The resulting matrix can be directly used with \code{PCAtools} or \code{ComplexHeatmap}
for further downstream explorative analysis e.g. unsupervised clustering
}
\examples{

mat <- matrix(sample.int(15, 9*100, TRUE), 9, 100) \%>\% as.data.frame()
mat <- mat \%>\%
tibble::rowid_to_column(var = "start") \%>\%
dplyr::mutate(end = start + 1000) \%>\%
dplyr::mutate(chr = "chr1") \%>\%
dplyr::select(chr, start, end, dplyr::everything())

get_variable_regions(enrichments_df = mat, num_regions = 50)
}
