% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomic_annotations.R
\name{merge_GR}
\alias{merge_GR}
\title{Merge genomic regions}
\usage{
merge_GR(x)
}
\arguments{
\item{x}{a character vector of bed file paths}
}
\value{
GRanges object
}
\description{
merge overlaping genomic regions from multiple peak/bed files
}
\examples{

## devtools::install_github("itsvenu/ALPSdata")

## bed files
chr21_bed_zip <- system.file("extdata", "chr21.bed.tar.gz", package = "ALPSdata", mustWork = TRUE)

untar(chr21_bed_zip, exdir = "~/Desktop/test_bed")
chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)
chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bed_path <- paste0("~/Desktop/test_bed/", chr21_data_table$bed_path)
x <- chr21_data_table$bed_path \%>\% as.character()

merge_GR(x = x)
}
