% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_browser_tracks.R
\name{plot_browser_tracks}
\alias{plot_browser_tracks}
\title{UCSC Genome browser like plots}
\usage{
plot_browser_tracks(data_table = NULL, gene_range = NULL,
  ref_gen = "hg38", cex.axis = 0.5, cex.title = 0.8, ...)
}
\arguments{
\item{data_table}{a dataframe that contains \code{bw_path}, \code{sample_id} and \code{color_code}. Tracks are colored according to \code{color_code}. Default \code{NULL}}

\item{gene_range}{genomic region for which browser-like plots needed in format \code{chr:start-end}. Default \code{NULL}}

\item{ref_gen}{reference genome, to get gene annotations, currently supports  \code{hg19} and \code{hg38}. Default, \code{hg38}}

\item{cex.axis}{axis label size, default \code{0.5}}

\item{cex.title}{axis title size, default \code{0.8}}

\item{...}{additional arguments to change the appearence of a plot. All arguments that can be passed to base R graphics are supported}
}
\value{
plot of genome browser tracks
}
\description{
Function to plot genome browser like plots given a very minimal information such as a set of bigwig files and a genomics region.
Tracks are arranged as they are in given input data.frame. Function uses highly customizable Gviz R/bioconductor package to plot browser like plots.
}
\examples{

## plot browser tracks
## devtools::install_github("itsvenu/ALPSdata")

chr21_bw_zip <- system.file("extdata", "chr21.bw.tar.gz", package = "ALPSdata", mustWork = TRUE)
untar(chr21_bw_zip, exdir = "./test_chr21")
chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)
chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bw_path <- paste0("./test_chr21/", chr21_data_table$bw_path)

set.seed(12345)
chr21_10samples <- chr21_data_table \%>\%
dplyr::sample_n(size = 10) \%>\%
dplyr::arrange(desc(group))

## gene_range
gene_range = "chr21:33025845-33129196"

plot_browser_tracks(data_table = chr21_10samples,
gene_range = gene_range, ref_gen = "hg38")
}
