% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Correlations among replicates/groups}
\usage{
plot_correlation(enrichments_df, log_transform = TRUE,
  method = "pearson", plot_type = "replicate_level", sample_metadata,
  ...)
}
\arguments{
\item{enrichments_df}{dataframe of enrichments, usually in the form of the output from function \code{\link{multiBigwig_summary}}, default \code{NULL}}

\item{log_transform}{logical, whether to log2 transform \code{enrichments_df}}

\item{method}{method to calculate correlation coefficient, one of \code{"pearson"} (default), \code{"spearman"} or \code{"kendall"}}

\item{plot_type}{whether to plot \code{replicate_level} correlations or \code{group_level} correlations. \code{replicate_level} plot represents the pairwise correlation values among all the samples/columns, where as the \code{group_level} plot is a paired plot of genomic regions after averaging of all samples in a group. Default \code{replicate_level}}

\item{sample_metadata}{a data.frame required if \code{plot_type = "group_level"}. The data.frame must contain columns \code{sample_id} and \code{group}}

\item{...}{additional arguments either to \code{corrplot::corrplot} or \code{GGally::ggpairs} depending on arg \code{plot_type}}
}
\value{
corrplot or ggplot2 object
}
\description{
Function to calculate correlations of ChIP/ATAC-seq enrichment among replicates/samples or groups. The function is compatible with the output of \code{\link{multiBigwig_summary}} or any custom data.frame with the similar format.
}
\examples{

## load example data
## devtools::install_github("itsvenu/ALPSdata")

## bigwig files
chr21_bw_zip <- system.file("extdata", "chr21.bw.tar.gz", package = "ALPSdata", mustWork = TRUE)

## untar into the current working dir under 'test_bw'
untar(chr21_bw_zip, exdir = "~/Desktop/test_bw")

## bed files
chr21_bed_zip <- system.file("extdata", "chr21.bed.tar.gz", package = "ALPSdata", mustWork = TRUE)

untar(chr21_bed_zip, exdir = "~/Desktop/test_bed")
chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)
chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bw_path <- paste0("~/Desktop/test_bw/", chr21_data_table$bw_path)
chr21_data_table$bed_path <- paste0("~/Desktop/test_bed/", chr21_data_table$bed_path)

chr21_data_table \%>\% head

enrichments_df <- multiBigwig_summary(data_table = chr21_data_table,
                                   summary_type = "mean",
                                   parallel = FALSE)

## subset to only ACCx samples
ACCx_all_replicates <- enrichments_df \%>\% dplyr::select(chr:ACCx_5)
ACCx_meta <- chr21_data_table \%>\% dplyr::filter(group == "ACCx")

## replicate_level correlation plot
plot_correlation(enrichments_df = ACCx_all_replicates, log_transform = TRUE, plot_type = "replicate_level", sample_metadata = ACCx_meta)

## group_level correlation plot
enrichments_df_grouplevel <- enrichments_df \%>\% dplyr::select(chr:CESC_5)
grouplevel_meta <- chr21_data_table \%>\% head(n = 20)
plot_correlation(enrichments_df = enrichments_df_grouplevel, log_transform = TRUE, plot_type = "group_level", sample_metadata = grouplevel_meta)
}
