% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genomic_annotations.R
\name{plot_genomic_annotations}
\alias{plot_genomic_annotations}
\title{Plot genomic annotations}
\usage{
plot_genomic_annotations(annotations_df = NULL, plot_type = "bar",
  col = NULL)
}
\arguments{
\item{annotations_df}{a data.frame of genomic annotations.
It can either be of one sample or of multiple samples as a data.frame}

\item{plot_type}{either \code{bar} plot or \code{heatmap},
default \code{bar}}

\item{col}{vector of colors for each feature in \code{annotations_df}.
If provided these colors are used, if not \code{pals::polychrome}
colors will be used. Default \code{NULL}}
}
\value{
\code{ggplot2} plot
}
\description{
plot the annotations of genomic regions
either as \code{stacked bar} or \code{heatmap}.
The function takes the output of \code{\link{get_genomic_annotations}}
directly or it is also compatible with a similar data.frame.
}
\examples{

## load example data
## load example data

chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)

## attach path to bw_path and bed_path
d_path <- dirname(chr21_data_table)

chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bw_path <- paste0(d_path, "/", chr21_data_table$bw_path)
chr21_data_table$bed_path <- paste0(d_path, "/", chr21_data_table$bed_path)

g_annotations <- get_genomic_annotations(data_table = chr21_data_table,
merge_level = "group_level")

plot_genomic_annotations(annotations_df = g_annotations, plot_type = "heatmap")
}
