% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_logo.R
\name{plot_motif_logo}
\alias{plot_motif_logo}
\title{Plot sequence motifs}
\usage{
plot_motif_logo(motif_path, database = NULL, plot_type = "bar")
}
\arguments{
\item{motif_path}{path to motif file, default \code{NULL}}

\item{database}{database name from which motif has taken, default \code{NULL}}

\item{plot_type}{either \code{bar} or \code{logo}, default \code{bar}}
}
\value{
\code{ggplot2} object
}
\description{
Function to plot transscription factor motif logos in two different styles, \code{bar} plot or \code{logo} plot.
It supports motif formats from different databases e.g. \code{JASPAR}, \code{MEME}, \code{TRANSFAC}, \code{HOMER} and \code{PFM}
}
\examples{

## examplr motif file paths
myc_meme <- system.file('extdata/motifs', 'MA0147.2.meme', package = 'ALPS', mustWork = TRUE)
myc_jaspar <- system.file('extdata/motifs', 'MA0147.2.jaspar', package = 'ALPS', mustWork = TRUE)
myc_transfac <- system.file('extdata/motifs', 'MA0147.2.transfac', package = 'ALPS', mustWork = TRUE)
myc_homer <- system.file('extdata/motifs', 'cmyc.homer', package = 'ALPS', mustWork = TRUE)
myc_pfm <- system.file('extdata/motifs', 'MA0147.2.pfm', package = 'ALPS', mustWork = TRUE)

## plot motifs
plot_motif_logo(motif_path = myc_homer, database = 'homer', plot_type = 'logo')
}
