% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_logo.R
\name{process_homer}
\alias{process_homer}
\title{Process homer format}
\usage{
process_homer(x)
}
\arguments{
\item{x}{path to homer format file}
}
\value{
data.frame
}
\description{
Process homer format
}
\examples{
myc_homer <- system.file("extdata/motifs", "cmyc.homer", package = "ALPS", mustWork = TRUE)
myc_df <- process_homer(x = myc_homer)
}
