% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_logo.R
\name{process_jaspar}
\alias{process_jaspar}
\title{Process jaspar format}
\usage{
process_jaspar(x)
}
\arguments{
\item{x}{path to jaspar format file}
}
\value{
data.frame
}
\description{
Process jaspar format
}
\examples{
myc_jaspar <- system.file('extdata/motifs', 'MA0147.2.jaspar', package = 'ALPS', mustWork = TRUE)
myc_df <- process_jaspar(x = myc_jaspar)
}
