% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_logo.R
\name{process_meme}
\alias{process_meme}
\title{Process meme format}
\usage{
process_meme(x)
}
\arguments{
\item{x}{path to meme format file}
}
\value{
data.frame
}
\description{
Process meme format
}
\examples{
myc_meme <- system.file('extdata/motifs', 'MA0147.2.meme', package = 'ALPS', mustWork = TRUE)
myc_df <- process_meme(x = myc_meme)
}
