% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_motif_logo.R
\name{process_transfac}
\alias{process_transfac}
\title{Process transfac format}
\usage{
process_transfac(x)
}
\arguments{
\item{x}{path to transfac format file}
}
\value{
data.frame
}
\description{
Process transfac format
}
\examples{
myc_transfac <- system.file('extdata/motifs', 'MA0147.2.transfac', package = 'ALPS', mustWork = TRUE)
my_df <- process_transfac(x = myc_transfac)
}
