% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cdf_df.R
\name{read.area}
\alias{read.area}
\title{Read global land fraction data}
\usage{
read.area(filename, tag = basename(filename), varname = "sftlf",
  latvar = "lat", lonvar = "lon")
}
\arguments{
\item{filename}{File name to read from}

\item{tag}{Identifier for the source of this dataset}

\item{varname}{Name of the variable to read}

\item{len}{Maximum length to read}
}
\description{
Read globally gridded time independent data from a netCDF file.
}
