% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cdf_df.R
\name{read.geo.aggregate}
\alias{read.geo.aggregate}
\title{Read globally averaged data}
\usage{
read.geo.aggregate(filename, len = NULL, tag = basename(filename),
  varname = "tas", timevar = "time")
}
\arguments{
\item{filename}{File name to read from}

\item{len}{Maximum length to read}

\item{tag}{Identifier for the source of this dataset}

\item{varname}{Name of the variable to read}

\item{timevar}{Name of the time variable}
}
\description{
Read a time series of globally averaged data from a netCDF file.
}
