% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cdf_df.R
\name{std_normalize}
\alias{std_normalize}
\title{Standardize data to zero mean and unit variance}
\usage{
std_normalize(df, stvars = c("global_value", "cell_value"))
}
\arguments{
\item{df}{A data frame of values to standardize}

\item{stvars}{List of names of variables to standardize in the df}
}
\value{
Data frame of standardized values.  The original means and standard
deviations will be added as attributes \code{means} and \code{stdevs}.
}
\description{
A single standardization is performed over the entire dataset, as
opposed to a bunch of group-wise standardizations (e.g., with the
groups being grid cells).  This preserves relationships between grid cells,
making it easier to interpret coefficients.
}
