% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALPS-package.r
\docType{package}
\name{ALPS-package}
\alias{ALPS}
\alias{ALPS-package}
\title{ALPS: AnaLysis routines for ePigenomicS data}
\description{
The package provides analysis and publication quality visualization routines for genome-wide epigenomics data such as histone modification or transcription factor ChIP-seq, ATAC-seq, DNase-seq etc. The functions in the package can be used with any type of data that can be represented with bigwig files at any resolution. The goal of the ALPS is to provide analysis tools for most downstream analysis without leaving the R environment and most tools in the package require a minimal input that can be prepared with basic R, unix or excel skills.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/itsvenu/ALPS}
  \item Report bugs at \url{https://github.com/itsvenu/ALPS/issues}
}

}
\keyword{internal}
