% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomic_annotations.R
\name{merge_GR}
\alias{merge_GR}
\title{Merge genomic regions}
\usage{
merge_GR(x)
}
\arguments{
\item{x}{a character vector of bed file paths}
}
\value{
GRanges object
}
\description{
merge overlaping genomic regions from multiple peak/bed files
}
\examples{

## load example data

chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)

## attach path to bw_path and bed_path
d_path <- dirname(chr21_data_table)

chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bw_path <- paste0(d_path, "/", chr21_data_table$bw_path)
chr21_data_table$bed_path <- paste0(d_path, "/", chr21_data_table$bed_path)

x <- as.character(chr21_data_table$bed_path)

merge_GR(x = x)
}
