% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBigwig_summary.R
\name{multiBigwig_summary}
\alias{multiBigwig_summary}
\title{Enrichments at genomics regions}
\usage{
multiBigwig_summary(data_table = NULL, summary_type = "mean",
  parallel = TRUE)
}
\arguments{
\item{data_table}{a dataframe that contains \code{bw_path}, \code{sample_id}.
\code{sample_id} ids will be used in the final result matrix}

\item{summary_type}{whether to calculate mean, median, min or max
for each genomic region in within consensus peak-set
from the bigwig files. Default is \code{mean}}

\item{parallel}{logical. Whether to parallelize the calculation process,
default is \code{TRUE}}
}
\value{
data.frame of enrichments within given genomic regions
}
\description{
\code{multiBigwig_summary} is a function to calculate
enrichments from a set of given bigwig files
and a set of genomics regions.
This function is similar to
\code{deeptools multiBigwigSummary} python package.
}
\examples{

## load example data

chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)

## attach path to bw_path and bed_path
d_path <- dirname(chr21_data_table)

chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bw_path <- paste0(d_path, "/", chr21_data_table$bw_path)
chr21_data_table$bed_path <- paste0(d_path, "/", chr21_data_table$bed_path)

enrichments <- multiBigwig_summary(data_table = chr21_data_table,
                                   summary_type = "mean",
                                   parallel = TRUE)
}
