% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_browser_tracks.R
\name{plot_browser_tracks}
\alias{plot_browser_tracks}
\title{UCSC Genome browser like plots}
\usage{
plot_browser_tracks(data_table = NULL, gene_range = NULL,
  ref_gen = "hg38", cex.axis = 0.5, cex.title = 0.8, ...)
}
\arguments{
\item{data_table}{a dataframe that contains \code{bw_path}, \code{sample_id} and \code{color_code}. Tracks are colored according to \code{color_code}. Default \code{NULL}}

\item{gene_range}{genomic region for which browser-like plots needed in format \code{chr:start-end}. Default \code{NULL}}

\item{ref_gen}{reference genome, to get gene annotations, currently supports  \code{hg19} and \code{hg38}. Default, \code{hg38}}

\item{cex.axis}{axis label size, default \code{0.5}}

\item{cex.title}{axis title size, default \code{0.8}}

\item{...}{additional arguments to change the appearence of a plot. All arguments that can be passed to base R graphics are supported}
}
\value{
plot of genome browser tracks
}
\description{
Function to plot genome browser like plots given a very minimal information such as a set of bigwig files and a genomics region.
Tracks are arranged as they are in given input data.frame. Function uses highly customizable Gviz R/bioconductor package to plot browser like plots.
}
\examples{

## load example data

chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)

## attach path to bw_path and bed_path
d_path <- dirname(chr21_data_table)

chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bw_path <- paste0(d_path, "/", chr21_data_table$bw_path)
chr21_data_table$bed_path <- paste0(d_path, "/", chr21_data_table$bed_path)

gene_range = "chr21:45643725-45942454"

plot_browser_tracks(data_table = chr21_data_table,
gene_range = gene_range, ref_gen = "hg38")
}
