% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Correlations among replicates/groups}
\usage{
plot_correlation(enrichments_df, log_transform = TRUE,
  method = "pearson", plot_type = "replicate_level", sample_metadata,
  ...)
}
\arguments{
\item{enrichments_df}{dataframe of enrichments, usually in the form of the output from function \code{\link{multiBigwig_summary}}, default \code{NULL}}

\item{log_transform}{logical, whether to log2 transform \code{enrichments_df}}

\item{method}{method to calculate correlation coefficient, one of \code{"pearson"} (default), \code{"spearman"} or \code{"kendall"}}

\item{plot_type}{whether to plot \code{replicate_level} correlations or \code{group_level} correlations. \code{replicate_level} plot represents the pairwise correlation values among all the samples/columns, where as the \code{group_level} plot is a paired plot of genomic regions after averaging of all samples in a group. Default \code{replicate_level}}

\item{sample_metadata}{a data.frame required if \code{plot_type = "group_level"}. The data.frame must contain columns \code{sample_id} and \code{group}}

\item{...}{additional arguments either to \code{corrplot::corrplot} or \code{GGally::ggpairs} depending on arg \code{plot_type}}
}
\value{
corrplot or ggplot2 object
}
\description{
Function to calculate correlations of ChIP/ATAC-seq enrichment among replicates/samples or groups. The function is compatible with the output of \code{\link{multiBigwig_summary}} or any custom data.frame with the similar format.
}
\examples{

## load example data
## load example data

chr21_data_table <- system.file("extdata/bw", "ALPS_example_datatable.txt", package = "ALPS", mustWork = TRUE)

## attach path to bw_path and bed_path
d_path <- dirname(chr21_data_table)

chr21_data_table <- read.delim(chr21_data_table, header = TRUE)
chr21_data_table$bw_path <- paste0(d_path, "/", chr21_data_table$bw_path)
chr21_data_table$bed_path <- paste0(d_path, "/", chr21_data_table$bed_path)

enrichments <- multiBigwig_summary(data_table = chr21_data_table,
                                   summary_type = "mean",
                                   parallel = TRUE)

## replicate_level correlation plot
plot_correlation(enrichments_df = enrichments,
log_transform = TRUE, plot_type = "replicate_level",
sample_metadata = chr21_data_table)

## group_level correlation plot
plot_correlation(enrichments_df = enrichments,
log_transform = TRUE, plot_type = "group_level",
sample_metadata = chr21_data_table)
}
